% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automobiles.R
\name{auto.delay}
\alias{auto.delay}
\title{Determine the intersection delay}
\usage{
auto.delay(link, int)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}

\item{dat}{Data.table of entire data set.}
}
\value{
Numeric value (decimal).
}
\description{
Determine the intersection delay
}
\examples{
auto.S_R(link, control)
}
