% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automobiles.R
\name{auto.S_f}
\alias{auto.S_f}
\title{Determine base free-flow speed}
\usage{
auto.S_f(link, int, dat)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}

\item{dat}{Data.table of entire data set.}
}
\value{
Numeric value in mi/hr.
}
\description{
Determine base free-flow speed
}
\examples{
auto.S_f(link, int, dat)
}
