% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automobiles.R
\name{auto.S_R}
\alias{auto.S_R}
\title{Determine segment running speed for automobiles}
\usage{
auto.S_R(link, int, dat)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}

\item{dat}{Data.table of entire data set.}
}
\value{
Numeric value in mi/hr.
}
\description{
Determine segment running speed for automobiles
}
\examples{
auto.S_R(link, control)
}
