% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automobiles.R
\name{auto.S_Tseg}
\alias{auto.S_Tseg}
\title{Determine the segment travel speed}
\usage{
auto.S_Tseg(link, int)
}
\arguments{
\item{link}{Data.table of link data.}

\item{int}{Data.table of subject intersection data.}
}
\value{
Numeric value (decimal).
}
\description{
Determine the segment travel speed
}
\examples{
auto.S_R(link, control)
}
