% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automobiles.R
\name{auto.satflow}
\alias{auto.satflow}
\title{Determine the adjusted saturation flow rate}
\usage{
auto.satflow(link, int)
}
\arguments{
\item{link}{Data.table of link data.}

\item{control}{String containing the boundary intersection control type
(Signalized", "AWSC - Stop", "TWSC - Stop", "Uncontrolled", "Yield")}
}
\value{
Numeric value in mi/hr.
}
\description{
Determine the adjusted saturation flow rate
}
\examples{
auto.satflow(link, int)
}
