% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automobiles.R
\name{auto.t_R}
\alias{auto.t_R}
\title{Determine the segment running time}
\usage{
auto.t_R(link, int, dat)
}
\arguments{
\item{link}{Data.table of link data.}

\item{int}{Data.table of subject intersection data.}
}
\value{
Numeric value (decimal).
}
\description{
Determine the segment running time
}
\examples{
auto.S_R(link, control)
}
