% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automobiles.R
\name{auto.VCratio}
\alias{auto.VCratio}
\title{Determine the volume to capacity ratio for critical intersection}
\usage{
auto.VCratio(link, int)
}
\arguments{
\item{link}{Data.table of link data.}

\item{int}{Data.table of subject intersection data.}
}
\value{
Numeric value (decimal).
}
\description{
Determine the volume to capacity ratio for critical intersection
}
\examples{
auto.VCratio(link, int)
}
