% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.d_2stageleft}
\alias{bike.d_2stageleft}
\title{Two-stage left turn bicycle delay}
\usage{
bike.d_2stageleft(int, dir)
}
\arguments{
\item{int}{Data.table of intersection data.}

\item{dir}{String with subject intersection approach being studied ("NB","SB","EB","WB").}
}
\value{
Numeric value in average seconds per bike.
}
\description{
Two-stage left turn bicycle delay
}
\examples{
bike.d_2stageleft(int, dir)
}
