% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.d_bd}
\alias{bike.d_bd}
\title{Average bicycle delay for intersection}
\usage{
bike.d_bd(int, dir)
}
\arguments{
\item{dir}{String with subject intersection approach being studied ("NB","SB","EB","WB")}

\item{link}{Data.table of link data.}
}
\value{
Numeric value, unitless.
}
\description{
Average bicycle delay for intersection
}
\examples{
bike.d_bd(int, dir)
}
