% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.d_bS}
\alias{bike.d_bS}
\title{Bicycle delay from signal and right turning vehicles}
\usage{
bike.d_bS(int, dir)
}
\arguments{
\item{int}{Data.table of intersection data.}

\item{dir}{String with subject intersection approach being studied ("NB","SB","EB","WB").}
}
\value{
Numeric value in average seconds per bike.
}
\description{
Bicycle delay from signal and right turning vehicles
}
\examples{
bike.d_bS(int, dir)
}
