% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.d_signal}
\alias{bike.d_signal}
\title{Bicycle control delay}
\usage{
bike.d_signal(link, int)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}
}
\value{
Numeric value in average seconds per bike.
}
\description{
Bicycle control delay
}
\examples{
bike.d_signal(int, dir)
}
