% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.d_twsc}
\alias{bike.d_twsc}
\title{Bicycle delay for uncontrolled intersections (e.g., Two-way Stop Controlled Intersections)}
\usage{
bike.d_twsc(link, int)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}
}
\value{
Numeric value in average seconds per bike.
}
\description{
Bicycle delay for uncontrolled intersections (e.g., Two-way Stop Controlled Intersections)
}
\examples{
bike.d_bS(int, dir)
}
