% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.F_w.link}
\alias{bike.F_w.link}
\title{Bicycle paved width factor (links)}
\usage{
bike.F_w.link(link, int)
}
\arguments{
\item{link}{Data.table of link data.}
}
\value{
Numeric value, unitless.
}
\description{
Bicycle paved width factor (links)
}
\examples{
bike.F_w.link(link)
}
