% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.I_int}
\alias{bike.I_int}
\title{Bicycle LOS score for intersections}
\usage{
bike.I_int(int, dir)
}
\arguments{
\item{int}{Data.table of intersection data.}

\item{dir}{String with subject intersection approach being studied ("NB","SB","EB","WB").}
}
\value{
A numeric LOS score, unitless.
}
\description{
Bicycle LOS score for intersections
}
\examples{
bike.I_int(int, dir)
}
