% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.I_link}
\alias{bike.I_link}
\title{Bicycle level of service score for links}
\usage{
bike.I_link(link, control)
}
\arguments{
\item{link}{Data.table of link data.}

\item{control}{String containing the boundary intersection control type
(Signalized", "AWSC - Stop", "TWSC - Stop", "Uncontrolled", "Yield").}
}
\value{
A numeric LOS score, unitless.
}
\description{
Bicycle level of service score for links
}
\examples{
bike.I_link(link, control)
}
