% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{MMLOS}
\alias{MMLOS}
\title{Calculate the Multimodal Level of Service}
\usage{
MMLOS(dat, revs = T)
}
\arguments{
\item{dat}{List containing data table for intersections and links. See \code{\link{loaddat}} for formatting.}

\item{revs}{Boolean where TRUE calculates LOS with proposed revisions, or FALSE using existing HCM 6th edition.}
}
\value{
A list containing a data tables for bicycle LOS and pedestrian LOS
}
\description{
Calculate the Multimodal Level of Service
}
\examples{
MMLOS(dat, T)
}
