% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{MMLOScalc}
\alias{MMLOScalc}
\title{Calculate the Multi-modal Level of Service}
\usage{
MMLOScalc(input_data, revs = T)
}
\arguments{
\item{input_data}{List containing data table for intersections and links. See \code{\link{loadMMLOS}} for formatting.}

\item{revs}{Boolean where TRUE calculates LOS with proposed revisions, or FALSE using existing HCM 6th edition.}
}
\value{
A list containing a data tables for bicycle LOS and pedestrian LOS
}
\description{
Calculate the Multi-modal Level of Service
}
\examples{
calcMMLOS(input_data, T)
}
