% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{MMLOSsave}
\alias{MMLOSsave}
\title{Save the output somewhere to CSV}
\usage{
MMLOSsave(dat, output_folder)
}
\arguments{
\item{dat}{The data you are saving}

\item{output_folder}{String containing the file output location.}
}
\description{
Save the output somewhere to CSV
}
\examples{
saveMMLOS(dat = los_norcal, output_folder = "./outputdata")
}
