% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles_og.R
\name{ogbike.F_s.link}
\alias{ogbike.F_s.link}
\title{Bicycle traffic speed factor (links)}
\usage{
ogbike.F_s.link(link, int, dat)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}

\item{dat}{Data.table of entire data set.}
}
\value{
Numeric value, unitless.
}
\description{
Bicycle traffic speed factor (links)
}
\examples{
ogbike.F_s.link(link)
}
