% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles_og.R
\name{ogbike.I_int}
\alias{ogbike.I_int}
\title{Bicycle LOS score for intersections}
\usage{
ogbike.I_int(int, dir)
}
\arguments{
\item{int}{Data.table of intersection data.}

\item{dir}{String with subject intersection approach being studied ("NB","SB","EB","WB")}
}
\value{
A numeric LOS score
}
\description{
Bicycle LOS score for intersections
}
\examples{
ogbike.I_int(int, dir)
}
