% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles_og.R
\name{ogbike.I_link}
\alias{ogbike.I_link}
\title{Bicycle level of service score for links}
\usage{
ogbike.I_link(link, int, dat)
}
\arguments{
\item{link}{Data.table of link data.}

\item{control}{String containing the boundary intersection control type (Signalized", "AWSC - Stop", "TWSC - Stop", "Uncontrolled", "Yield")}
}
\value{
A numeric LOS score
}
\description{
Bicycle level of service score for links
}
\examples{
ogbike.I_link(link, control)
}
