% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles_og.R
\name{ogbike.I_seg}
\alias{ogbike.I_seg}
\title{Bicycle LOS score for segment}
\usage{
ogbike.I_seg(link, int, dat)
}
\arguments{
\item{link}{Data.table of link data.}

\item{int}{Data.table of intersection data.}
}
\value{
A data.table with numeric and letter grade LOS scores
}
\description{
Bicycle LOS score for segment
}
\examples{
ogbike.I_seg(link, int)
}
