% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians_og.R
\name{ogped.d_twsc}
\alias{ogped.d_twsc}
\title{Revised pedestrian delay for uncontrolled (e.g., TWSC)}
\usage{
ogped.d_twsc(int, dir)
}
\arguments{
\item{int}{Data.table of intersection data.}

\item{dir}{String with subject intersection approach being studied ("NB","SB","EB","WB")}
}
\value{
Numeric value, unitless.
}
\description{
Revised pedestrian delay for uncontrolled (e.g., TWSC)
}
\examples{
ogped.d_twsc(int, dir)
}
