% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians_og.R
\name{ogped.F_w.link}
\alias{ogped.F_w.link}
\title{Pedestrian paved width factor (links)}
\usage{
ogped.F_w.link(link)
}
\arguments{
\item{link}{Data.table of link data.}
}
\value{
Numeric value, unitless.
}
\description{
Pedestrian paved width factor (links)
}
\examples{
ogped.F_w.link(link)
}
