% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians_og.R
\name{ogped.I_int}
\alias{ogped.I_int}
\title{Pedestrian LOS score for intersections}
\usage{
ogped.I_int(int, dir)
}
\arguments{
\item{int}{Data.table of boundary intersection data.
(Signalized", "AWSC - Stop", "TWSC - Stop", "Uncontrolled", "Yield")}

\item{link}{Data.table of link data.}
}
\value{
A numeric LOS score
}
\description{
Pedestrian LOS score for intersections
}
\examples{
ped.I_int(link, int)
}
