% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians_og.R
\name{ogped.I_link}
\alias{ogped.I_link}
\title{Pedestrian level of service score for links}
\usage{
ogped.I_link(link, int, dat)
}
\arguments{
\item{link}{Data.table of link data.}

\item{int}{Data.table of intersection data.}
}
\value{
Numeric value, unitless.
}
\description{
Pedestrian level of service score for links
}
\examples{
ogped.I_link(int, dir)
}
