% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians.R
\name{ped.d_signal}
\alias{ped.d_signal}
\title{Pedestrian control delay from signal}
\usage{
ped.d_signal(int, dir)
}
\arguments{
\item{int}{Data.table of intersection data.}

\item{dir}{String with subject intersection approach being studied ("NB","SB","EB","WB")}
}
\value{
Numeric value, unitless.
}
\description{
Pedestrian control delay from signal
}
\examples{
ped.d_signal(int, dir)
}
