% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians.R
\name{ped.d_twsc}
\alias{ped.d_twsc}
\title{Revised pedestrian delay for uncontrolled (e.g., TWSC)}
\usage{
ped.d_twsc(int, dir)
}
\arguments{
\item{int}{Data.table of intersection data.}

\item{dir}{String with subject intersection approach being studied ("NB","SB","EB","WB")}
}
\value{
Numeric value, unitless.
}
\description{
Revised pedestrian delay for uncontrolled (e.g., TWSC)
}
\examples{
ped.d_twsc(int, dir)
}
