% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians.R
\name{ped.F_s.link}
\alias{ped.F_s.link}
\title{Pedestrian traffic speed factor (links)}
\usage{
ped.F_s.link(link, int)
}
\arguments{
\item{link}{Data.table of link data.}

\item{int}{Data.table of intersection data.}
}
\value{
Numeric value, unitless.
}
\description{
Pedestrian traffic speed factor (links)
}
\examples{
ped.F_s.link(int, dir)
}
