% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians.R
\name{ped.F_w.link}
\alias{ped.F_w.link}
\title{Pedestrian paved width factor (links)}
\usage{
ped.F_w.link(link)
}
\arguments{
\item{link}{Data.table of link data.}
}
\value{
Numeric value, unitless.
}
\description{
Pedestrian paved width factor (links)
}
\examples{
ped.F_w.link(link)
}
