% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians.R
\name{ped.I_mx}
\alias{ped.I_mx}
\title{Pedestrian level of service score for midsegment crossings}
\usage{
ped.I_mx(link, int)
}
\arguments{
\item{link}{Data.table of link data.}

\item{int}{Data.table of boundary intersection data.
(Signalized", "AWSC - Stop", "TWSC - Stop", "Uncontrolled", "Yield")}
}
\value{
A numeric LOS score
}
\description{
Pedestrian level of service score for midsegment crossings
}
\examples{
ped.I_mx(link, int)
}
