% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{score2LOS}
\alias{score2LOS}
\title{Convert LOS score to letter grade}
\usage{
score2LOS(score, mode)
}
\arguments{
\item{score}{Numeric LOS score}

\item{mode}{String for mode being converted. Bike or Pedestrian (future version to include Automobile and Transit)}
}
\value{
A character string for LOS grade.
}
\description{
Convert LOS score to letter grade
}
\examples{
score2LOS(score, mode)
}
