% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automobiles.R
\name{auto.S_R}
\alias{auto.S_R}
\title{Determine segment running speed for automobiles}
\usage{
auto.S_R(link, control)
}
\arguments{
\item{link}{Data.table of link data.}

\item{control}{String containing the boundary intersection control type
(Signalized", "AWSC - Stop", "TWSC - Stop", "Uncontrolled", "Yield")}
}
\value{
Numeric value in mi/hr.
}
\description{
Determine segment running speed for automobiles
}
\examples{
auto.S_R(link, control)
}
