% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automobiles.R
\name{auto.VCratio}
\alias{auto.VCratio}
\title{Determine the volume to capacity ratio for critical intersection}
\usage{
auto.VCratio(link, int)
}
\arguments{
\item{link}{Data.table of link data.}

\item{control}{String containing the boundary intersection control type
(Signalized", "AWSC - Stop", "TWSC - Stop", "Uncontrolled", "Yield")}
}
\value{
Numeric value in mi/hr.
}
\description{
Determine the volume to capacity ratio for critical intersection
}
\examples{
auto.S_R(link, control)
}
