% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.d_2stageleft}
\alias{bike.d_2stageleft}
\title{Two-stage left turn bicycle delay}
\usage{
bike.d_2stageleft(link, int)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}
}
\value{
Numeric value in average seconds per bike.
}
\description{
Two-stage left turn bicycle delay
}
\examples{
bike.d_2stageleft(int, dir)
}
