% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.d_bS}
\alias{bike.d_bS}
\title{Bicycle delay from signal and right turning vehicles}
\usage{
bike.d_bS(link, int)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}
}
\value{
Numeric value in average seconds per bike.
}
\description{
Bicycle delay from signal and right turning vehicles
}
\examples{
bike.d_bS(int, dir)
}
