% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.F_s.link}
\alias{bike.F_s.link}
\title{Bicycle traffic speed factor (links)}
\usage{
bike.F_s.link(link, control)
}
\arguments{
\item{link}{Data.table of link data.}

\item{control}{String containing the boundary intersection control type
(Signalized", "AWSC - Stop", "TWSC - Stop", "Uncontrolled", "Yield")}
}
\value{
Numeric value, unitless.
}
\description{
Bicycle traffic speed factor (links)
}
\examples{
bike.F_s.link(link)
}
