% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.I_int}
\alias{bike.I_int}
\title{Bicycle LOS score for intersections}
\usage{
bike.I_int(link, int, dat)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}
}
\value{
A numeric LOS score, unitless.
}
\description{
Bicycle LOS score for intersections
}
\examples{
bike.I_int(int, dir)
}
