% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.I_link}
\alias{bike.I_link}
\title{Bicycle level of service score for links}
\usage{
bike.I_link(link, int, dat)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}
}
\value{
A numeric LOS score, unitless.
}
\description{
Bicycle level of service score for links
}
\examples{
bike.I_link(link, control)
}
