% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles.R
\name{bike.I_seg}
\alias{bike.I_seg}
\title{Bicycle LOS score for segment}
\usage{
bike.I_seg(link, int)
}
\arguments{
\item{link}{Data.table of link data.}

\item{int}{Data.table of intersection data.}
}
\value{
A data.table with numeric and letter grade LOS scores.
}
\description{
Bicycle LOS score for segment
}
\examples{
bike.I_seg(link, int)
}
