% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles_og.R
\name{ogbike.F_w.link}
\alias{ogbike.F_w.link}
\title{Bicycle paved width factor (links)}
\usage{
ogbike.F_w.link(link, int)
}
\arguments{
\item{link}{Data.table of link data.}

\item{int}{Data.table of subject intersection data.}
}
\value{
Numeric value, unitless.
}
\description{
Bicycle paved width factor (links)
}
\examples{
ogbike.F_w.link(link)
}
