% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycles_og.R
\name{ogbike.I_int}
\alias{ogbike.I_int}
\title{Bicycle LOS score for intersections}
\usage{
ogbike.I_int(link, int)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}

\item{dat}{Data.table of entire data set.}
}
\value{
A numeric LOS score
}
\description{
Bicycle LOS score for intersections
}
\examples{
ogbike.I_int(int, dir)
}
