% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians_og.R
\name{ogped.d_pd}
\alias{ogped.d_pd}
\title{Pedestrian control delay}
\usage{
ogped.d_pd(int, dir)
}
\arguments{
\item{dir}{String with subject intersection approach being studied ("NB","SB","EB","WB")}

\item{link}{Data.table of subject link data.}
}
\value{
Numeric value, unitless.
}
\description{
Pedestrian control delay
}
\examples{
ogped.d_pd(int, dir)
}
