% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians_og.R
\name{ogped.F_cd}
\alias{ogped.F_cd}
\title{Pedestrian level of service score for mid-segment crossings}
\usage{
ogped.F_cd(link, int, Ilink, Iint)
}
\arguments{
\item{link}{Data.table of link data.}

\item{Ilink}{Numeric LOS score for links.}

\item{Iint}{Numeric LOS score for intersections.}

\item{dir}{String with subject intersection approach being studied ("NB","SB","EB","WB")
(Signalized", "AWSC - Stop", "TWSC - Stop", "Uncontrolled", "Yield")}
}
\value{
A numeric LOS score
}
\description{
Pedestrian level of service score for mid-segment crossings
}
\examples{
ogped.F_cd(link, int, Ilink, Iint)
}
