% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians_og.R
\name{ogped.I_int}
\alias{ogped.I_int}
\title{Pedestrian LOS score for intersections}
\usage{
ogped.I_int(link, int, dat)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}

\item{dat}{Data.table of entire data set.}
}
\value{
A numeric LOS score
}
\description{
Pedestrian LOS score for intersections
}
\examples{
ped.I_int(link, int)
}
