% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians.R
\name{ped.d_pd}
\alias{ped.d_pd}
\title{Pedestrian control delay}
\usage{
ped.d_pd(int, dir)
}
\arguments{
\item{int}{Data.table of subject intersection data.}

\item{link}{Data.table of subject link data.}
}
\value{
Numeric value, unitless.
}
\description{
Pedestrian control delay
}
\examples{
ped.d_pd(int, dir)
}
