% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians.R
\name{ped.I_int}
\alias{ped.I_int}
\title{Pedestrian LOS score for intersections}
\usage{
ped.I_int(link, int)
}
\arguments{
\item{link}{Data.table of link data.}

\item{int}{Data.table of boundary intersection data.
(Signalized", "AWSC - Stop", "TWSC - Stop", "Uncontrolled", "Yield")}
}
\value{
A numeric LOS score
}
\description{
Pedestrian LOS score for intersections
}
\examples{
ped.I_int(link, int)
}
