% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrians.R
\name{ped.I_link}
\alias{ped.I_link}
\title{Pedestrian level of service score for links}
\usage{
ped.I_link(link, int, dat)
}
\arguments{
\item{link}{Data.table of subject link data.}

\item{int}{Data.table of subject intersection data.}

\item{dat}{Data.table of entire data set.}
}
\value{
Numeric value, unitless.
}
\description{
Pedestrian level of service score for links
}
\examples{
ped.I_link(int, dir)
}
